refactor: :art: general rules used in dto input to length validation
