+++ src/modules/session/domain/constants/general-rules.ts
@@ -1,0 +1,42 @@
+export const SESSION = {
+  SESSION_ID: {
+    MAX_LENGTH: 255,
+    MIN_LENGTH: 1,
+  },
+  SESSION_TYPE: {
+    MAX_LENGTH: 50,
+    MIN_LENGTH: 1,
+  },
+  SESSION_DURATION: {
+    MAX_LENGTH: 50,
+    MIN_LENGTH: 1,
+  },
+  SESSION_CLOSED_TYPE: {
+    MAX_LENGTH: 50,
+    MIN_LENGTH: 1,
+  },
+  TOKEN: {
+    MAX_LENGTH: 1000,
+    MIN_LENGTH: 1,
+  },
+  IP_ADDRESS: {
+    MAX_LENGTH: 100,
+    MIN_LENGTH: 1,
+  },
+  REFRESH_TOKEN: {
+    MAX_LENGTH: 255,
+    MIN_LENGTH: 1,
+  },
+  USER_AGENT: {
+    MAX_LENGTH: 255,
+    MIN_LENGTH: 1,
+  },
+  ORIGIN: {
+    MAX_LENGTH: 100,
+    MIN_LENGTH: 1,
+  },
+  LOCATION: {
+    MAX_LENGTH: 100,
+    MIN_LENGTH: 1,
+  },
+};
