+++ src/modules/session/domain/constants/injection-tokens.ts
@@ -1,0 +1,1 @@
+export const USER_REPOSITORY = Symbol('USER_REPOSITORY');
