+++ src/modules/user/domain/constants/general-rules.ts
@@ -59,48 +59,12 @@ export const MIN_ACCOUNT_EMAIL_LENGTH = 5;
 export const MAX_ACCOUNT_PASSWORD_LENGTH = 100;
 export const MIN_ACCOUNT_PASSWORD_LENGTH = 5;
 
-// TODO: Use in value objects
-// TODO: Use in dtos
-// TODO: Use in entities
-export const SESSION = {
-  SESSION_ID: {
-    MAX_LENGTH: 255,
-    MIN_LENGTH: 1,
-  },
-  SESSION_TYPE: {
-    MAX_LENGTH: 50,
-    MIN_LENGTH: 1,
-  },
-  SESSION_DURATION: {
-    MAX_LENGTH: 50,
-    MIN_LENGTH: 1,
-  },
-  SESSION_CLOSED_TYPE: {
-    MAX_LENGTH: 50,
-    MIN_LENGTH: 1,
-  },
-  TOKEN: {
-    MAX_LENGTH: 1000,
-    MIN_LENGTH: 1,
-  },
-  IP_ADDRESS: {
-    MAX_LENGTH: 100,
-    MIN_LENGTH: 1,
-  },
-  REFRESH_TOKEN: {
-    MAX_LENGTH: 255,
-    MIN_LENGTH: 1,
-  },
-  USER_AGENT: {
-    MAX_LENGTH: 255,
-    MIN_LENGTH: 1,
-  },
-  ORIGIN: {
-    MAX_LENGTH: 100,
-    MIN_LENGTH: 1,
-  },
-  LOCATION: {
-    MAX_LENGTH: 100,
-    MIN_LENGTH: 1,
-  },
-};
+export enum UserStatusEnum {
+  ACTIVE = 'Active',
+  INACTIVE = 'Inactive',
+  PENDING = 'Pending',
+  SUSPENDED = 'Suspended',
+  ARCHIVED = 'Archived',
+  BANNED = 'Banned',
+  UNDER_REVIEW = 'UnderReview',
+}
