+++ tsconfig.json
@@ -42,6 +42,12 @@
       "@user/*": [
         "src/modules/user/*"
       ],
+      "@session": [
+        "src/modules/session/"
+      ],
+      "@session/*": [
+        "src/modules/session/*"
+      ],
       "@role": [
         "src/modules/role/"
       ],
